Refactor for MVC achitecture
