+++ server.js
@@ -1,5 +1,6 @@
 const dotenv = require('dotenv');
 const mongoose = require('mongoose');
+const app = require('./app');
 
 dotenv.config({ path: './config.env' });
 
@@ -17,25 +18,6 @@ mongoose
   })
   .then(() => console.log('MongoDB connection successful'));
 
-const tourSchema = new mongoose.Schema({
-  name: {
-    type: String,
-    required: [true, 'A tour must have a name'],
-    unique: true,
-  },
-  rating: {
-    type: Number,
-    default: 4.5,
-  },
-  price: {
-    type: Number,
-    required: [true, 'A tour must have a price'],
-  },
-});
-
-const Tour = mongoose.model('Tour', tourSchema);
-
-const app = require('./app');
 // RUN SERVER
 const port = process.env.PORT || 4000;
 app.listen(port, () => {
